requirements.txt
# --------- pytorch --------- #
lightning[extra]==2.0.4
torchmetrics==0.11.4
timm==0.9.2
tiktoken==0.4.0

# --------- model zoo -------#
transformers==4.30.1

# ------- tensor utils ------#
einops==0.6.1

# --------- hydra --------- #
hydra-core==1.3.2
hydra-colorlog==1.2.0
hydra-joblib-launcher==1.2.0
hydra-optuna-sweeper==1.2.0

# --------- loggers --------- #
# wandb
# neptune-client
# mlflow
# comet-ml
aim==3.17.5  # no lower than 3.16.2, see https://github.com/aimhubio/aim/issues/2550

# --------- deployment ------ #
starlette<0.28.0
fastapi==0.98.0
gradio==3.36.1

# --------- others --------- #
dvc==3.2.1
pyrootutils==1.0.4     # standardizing the project root setup
pre-commit==3.3.2      # hooks for applying linters on commit
rich==13.4.2            # beautiful text formatting in terminal
pytest==7.3.2          # tests
black==23.3.0
# sh            # for running bash commands in some tests (linux/macos only)